[STEP1] 오탈자 수정
