+++ WeatherForecast/WeatherForecast/WeatherDetail/Models/Structs/WeatherDetailInfo.swift
@@ -6,20 +6,6 @@
 //
 
 import Foundation
-protocol WeatherDetailInfoCoordinator {
-    var sunrise: String { get }
-    var sunset: String { get }
-    func getWeather() -> String
-    func getDescription() -> String
-    func getTemp() -> String
-    func getTempMax() -> String
-    func getTempMin() -> String
-    func getFeelsLike() -> String
-    func getTempUnitExpression() -> String
-    func getPop() -> String
-    func getHumidity() -> String
-    func getIconName() -> String
-}
 
 struct WeatherDetailInfo: WeatherDetailInfoCoordinator {
     //MARK: - Properties
